Add the project and task modules
