+++ src/project.js
@@ -1,11 +1,20 @@
-const projectInput = document.querySelector('#project');
-const allProjects = document.querySelector('.added')
-const buttonProject = document.querySelector('.btn-project')
-
-const addProject=()=>{
-  const project = document.createElement('li');
-  project.innerHTML = projectInput.value
-  allProjects.appendChild(project)
+function ProjectList(name,allprojects =[]){
+  this.name=name
+  this.allprojects=allprojects
 }
-buttonProject.addEventListener('click',addProject)
-export default addProject
+
+ProjectList.prototype.addTask=(task)=>{
+  this.allprojects.push(task)
+}
+
+let projects = []
+const createProject = ()=>{
+  const newProject = projectList(name, allprojects)
+  projects.push(newProject)
+}
+
+const deleteProject= (index) =>{
+  projects.splice(index,1)
+} 
+
+export default {addTask,createProject,deleteProject}
